refactored tests to bring more granularity as eventually some of them require different implementations
